+++ example_app/server/jest.config.js
@@ -4,4 +4,7 @@ module.exports = {
       '^.+\\.(ts|tsx)?$': 'ts-jest',
       '^.+\\.(js|jsx)$': 'babel-jest',
     },
+    moduleNameMapper: {
+      '^@fullcircle/harness/(.*)$': '<rootDir>/../../packages/harness/src/$1',
+    },
   };
