+++ example_app/server/tsconfig.json
@@ -14,5 +14,8 @@
     "strict": true,
     "skipLibCheck": true,
     "outDir": "dist",
+    "paths": {
+      "@fullcircle/harness/*": ["../../packages/harness/src/*"]
+    }
   }
 }
