+++ packages/harness/package.json
@@ -4,8 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "npm run recorder",
-    "recorder": "npm start --prefix packages/recorder",
+    "start": "ts-node src/index.ts",
+    "dev": "ts-node-dev src/index.ts",
     "test": "jest && npm test --prefix example_app/server"
   },
   "keywords": [],
@@ -26,5 +26,10 @@
     "ts-node": "^10.9.1",
     "ts-node-dev": "^2.0.0",
     "typescript": "^5.2.2"
+  },
+  "dependencies": {
+    "express": "^4.18.2",
+    "express-async-errors": "^3.1.1",
+    "express-http-proxy": "^2.0.0"
   }
 }
