+++ packages/recorder/session_recording.ts/sessions_manager.ts
