+++ packages/recorder/src/express_app.ts
@@ -1,9 +1,9 @@
 import express from 'express';
 
-import {initFullCircleApiRouter} from './controllers/fullcircle_api_controller';
+import {initFullCircleApiRouter} from '../controllers/fullcircle_api_controller';
 
 // import {initProxyRouter} from './controllers/proxy_controller';
-import {initProxyRouter} from './controllers/with_express_http_proxy_library';
+import {initProxyRouter} from '../controllers/with_express_http_proxy_library';
 
 import {AppDependencies} from './types';
